+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.6] - 2026-01-30
+
+### Added
+
+- Font size adjustment for article list view
+  - New setting in General settings with -2 to +2 offset
+  - Affects article titles in both desktop and mobile layouts
+
 ## [0.4.5] - 2026-01-30
 
 ### Added